backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
reportlab
jinja2
requests
python-multipart
